Revert "Revert "Arquivos no formato padrão""
This reverts commit 39448b11682cfc3cd1cb0215eae7cbea107fa195.
