+++ app/nbproject/project.properties
@@ -74,9 +74,9 @@ javadoc.windowtitle=
 javafx.application.implementation.version=1.0
 javafx.binarycss=false
 javafx.classpath.extension=\
-    ${platforms.JDK_8.home}/jre/lib/javaws.jar:\
-    ${platforms.JDK_8.home}/jre/lib/deploy.jar:\
-    ${platforms.JDK_8.home}/jre/lib/plugin.jar
+    ${platforms.JDK_8u231.home}/jre/lib/javaws.jar:\
+    ${platforms.JDK_8u231.home}/jre/lib/deploy.jar:\
+    ${platforms.JDK_8u231.home}/jre/lib/plugin.jar
 javafx.deploy.adddesktopshortcut=false
 javafx.deploy.addstartmenushortcut=false
 javafx.deploy.allowoffline=true
@@ -95,7 +95,7 @@ javafx.enable.concurrent.external.runs=false
 javafx.enabled=true
 javafx.fallback.class=com.javafx.main.NoJavaFXFallback
 # Main class for JavaFX
-javafx.main.class=app.MainApp
+javafx.main.class=app.inicio.MainApp
 javafx.preloader.class=
 # This project does not use Preloader
 javafx.preloader.enabled=false
@@ -124,7 +124,7 @@ manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
 native.bundling.enabled=false
-platform.active=JDK_8
+platform.active=JDK_8u231
 run.classpath=\
     ${dist.jar}:\
     ${javac.classpath}:\
