+++ app/src/java/app/diario/telaloginadm/FXMLControllerTelaLoginAdm.java
@@ -1,4 +1,4 @@
-package app;
+package app.diario.telaloginadm;
 
 import java.net.URL;
 import java.util.ResourceBundle;
@@ -6,15 +6,15 @@ import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.control.Label;
 
-public class FXMLController implements Initializable {
+public class FXMLControllerTelaLoginAdm implements Initializable {
     
     @FXML
-    private Label label;
+//    private Label label;
     
     @Override
     public void initialize(URL url, ResourceBundle rb) {
         String javaVersion = System.getProperty("java.version");
         String javafxVersion = System.getProperty("javafx.version");
-        label.setText("Hello, JavaFX " + javafxVersion + "\nRunning on Java " + javaVersion + ".");
+//        label.setText("Hello, JavaFX " + javafxVersion + "\nRunning on Java " + javaVersion + ".");
     }    
 }
