+++ app/src/java/app/diario/telaloginadm/MainTelaLoginAdm.java
@@ -1,4 +1,4 @@
-package app;
+package app.diario.telaloginadm;
 
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
@@ -7,17 +7,18 @@ import javafx.scene.Scene;
 import javafx.stage.Stage;
 
 
-public class MainApp extends Application {
+public class MainTelaLoginAdm extends Application {
 
     @Override
     public void start(Stage stage) throws Exception {
-        Parent root = FXMLLoader.load(getClass().getResource("scene.fxml"));
+        Parent root = FXMLLoader.load(getClass().getResource("FXMLTelaLoginAdm.fxml"));
         
         Scene scene = new Scene(root);
-        scene.getStylesheets().add(getClass().getResource("styles.css").toExternalForm());
+        scene.getStylesheets().add(getClass().getResource("EstiloTelaLoginAdm.css").toExternalForm());
         
-        stage.setTitle("JavaFX and JDK8");
+        stage.setTitle("Sistema Acadêmico");
         stage.setScene(scene);
+        stage.setResizable(false);
         stage.show();
     }
