--- app/src/resources/app/diario/cursos/Editar.fxml
@@ -1,45 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.text.*?>
-<?import java.lang.*?>
-<?import java.util.*?>
-<?import javafx.scene.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-
-<AnchorPane id="AnchorPane" prefHeight="233.0" prefWidth="676.0" styleClass="fonte" stylesheets="@../Principal/style.css" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="app.diario.cursos.ControllerEditar">
-    <children>
-      <Label fx:id="idDeptoLabel" layoutX="139.0" layoutY="66.0" minHeight="16" minWidth="69" text="   ID-DEPTO" textAlignment="CENTER" textFill="#5dbaa7">
-         <font>
-            <Font name="System Bold" size="12.0" />
-         </font>
-      </Label>
-      <Label fx:id="nomeLabel" layoutX="131.0" layoutY="155.0" minHeight="16" minWidth="69" text="    NOME" textAlignment="CENTER" textFill="#5dbaa7">
-         <font>
-            <Font name="System Bold" size="12.0" />
-         </font>
-      </Label>
-      <Label fx:id="horasLabel" layoutX="268.0" layoutY="66.0" minHeight="16" minWidth="69" text="HORAS-TOTAL" textAlignment="CENTER" textFill="#5dbaa7">
-         <font>
-            <Font name="System Bold" size="12.0" />
-         </font>
-      </Label>
-      <Label fx:id="modalidadeLabel" layoutX="313.0" layoutY="155.0" minHeight="16" minWidth="69" text="MODALIDADE" textAlignment="CENTER" textFill="#5dbaa7">
-         <font>
-            <Font name="System Bold" size="12.0" />
-         </font>
-      </Label>
-      <Label layoutX="14.0" layoutY="23.0" text="Editar dados do curso:" textFill="#377256">
-         <font>
-            <Font size="29.0" />
-         </font>
-      </Label>
-      <Button fx:id="botaoCancelar" layoutX="469.0" layoutY="181.0" mnemonicParsing="false" onAction="#acaoCancelar" style="-fx-background-color: #E53935;" text="CANCELAR" textFill="WHITE" />
-      <Button fx:id="botaoSalvar" layoutX="558.0" layoutY="181.0" mnemonicParsing="false" onAction="#acaoSalvar" prefWidth="75.0" style="-fx-background-color: #29a58c;" text="SALVAR" textFill="WHITE" />
-      <TextField fx:id="idDeptoInput" layoutX="123.0" layoutY="94.0" prefWidth="100.0" style="-fx-background-color: #dddddd;" />
-      <TextField fx:id="horasInput" layoutX="260.0" layoutY="94.0" prefWidth="100.0" style="-fx-background-color: #dddddd;" />
-      <TextField fx:id="nomeInput" layoutX="91.0" layoutY="181.0" prefWidth="150.0" style="-fx-background-color: #dddddd;" />
-      <TextField fx:id="modalidadeInput" layoutX="279.0" layoutY="181.0" prefWidth="150.0" style="-fx-background-color: #dddddd;" />
-      <Label fx:id="errorLabel" layoutX="429.0" layoutY="155.0" styleClass="fonte" textFill="#e53935" />
-    </children>
-</AnchorPane>
