--- app/src/resources/app/diario/cursos/style.css
@@ -1,71 +1,0 @@
-@import url('https://fonts.googleapis.com/css?family=Montserrat&display=swap');
-
-table-view, .tree-table-view {
-    /* Constants used throughout the tableview. */
-    -fx-table-header-border-color: transparent;
-    -fx-table-cell-border-color: -fx-box-border; /* Horizontal Lines*/
-    -fx-background-color: transparent;
-}
-
-/* The column header row is made up of a number of column-header, one for each
-   TableColumn, and a 'filler' area that extends from the right-most column
-   to the edge of the tableview, or up to the 'column control' button. */
-.table-view .filler, .tree-table-view .filler, .table-view .column-header,
-.tree-table-view .column-header {
-    -fx-size: 65;
-    -fx-border-style: null;
-    -fx-border-width: 0 0 2 0;
-    -fx-background-color: transparent;
-    
-}
-
-.table-view .show-hide-columns-button, .tree-table-view .show-hide-columns-button
-{
-    -fx-background-color: transparent;
-}
-
-.table-view .column-header .label, .table-view .filler .label,
-.table-view .column-drag-header .label, .tree-table-view .column-header .label,
-.tree-table-view .filler .label, .tree-table-view .column-drag-header .label
-{
-    -fx-alignment: CENTER_LEFT;
-    -fx-text-fill: #377256;
-}
-
-.table-view .column-header-background, .tree-table-view .column-header-background
-{
-    -fx-background-color: transparent;
-    
-    
-}
-
-.table-row-cell, .tree-table-row-cell {
-    -fx-cell-size: 40px;
-}
-
-.table-cell {
-    -fx-border-color: transparent; /* Vertical Lines*/
-    -fx-border-width: 1;
-}
-.table-view{
-    -fx-border-color: transparent;
-    -fx-alignment: CENTER_LEFT;
-    
-}
-#deleta{
-    -fx-background-color: #E53935;
-    -fx-text-fill: white;
-    
-}
-#edita{
-    -fx-background-color: #2C46A5;
-    -fx-text-fill: white;
-
-}
-.coluna{
-    -fx-font-family: 'Montserrat',sans-serif;
-}
-
-.fonte{
-    -fx-font-family: 'Montserrat',sans-serif;
-}
