+++ app/src/resources/app/inicio/EstiloInicio.css
@@ -1,0 +1,25 @@
+@import url('https://fonts.googleapis.com/css?family=Montserrat&display=swap');
+
+#lblSistemaAcademico {
+    -fx-background-color: #29a58c;
+}
+
+#btnEntraSistemaDiario {
+    -fx-pref-width: 250;
+    -fx-pref-heigth: 50;
+    /*-fx-font-family: 'Montserrat', sans-serif;*/
+}
+
+#btnEntraSistemaBiblioteca {
+    -fx-pref-width: 250;
+    -fx-pref-heigth: 50;
+    /*-fx-font-family: 'Montserrat', sans-serif;*/
+}
+
+.backgroundBtn {
+    -fx-background-color: #29A58C;
+}
+
+.fontFamily {
+    -fx-font-family: 'Montserrat', sans-serif;
+}
