--- app/src/resources/app/styles.css
@@ -1,1 +1,0 @@
-
