--- diario/Cursos/nbproject/configs/Run_as_WebStart.properties
@@ -1,2 +1,0 @@
-# Do not modify this property in this configuration. It can be re-generated.
-$label=Run as WebStart
