--- diario/Cursos/nbproject/genfiles.properties
@@ -1,8 +1,0 @@
-build.xml.data.CRC32=fbcf2972
-build.xml.script.CRC32=0a444dff
-build.xml.stylesheet.CRC32=8064a381@1.80.1.48
-# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
-# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
-nbproject/build-impl.xml.data.CRC32=5a3c2ded
-nbproject/build-impl.xml.script.CRC32=8904f8fb
-nbproject/build-impl.xml.stylesheet.CRC32=3a2fa800@1.90.1.48
