--- diario/Cursos/nbproject/project.properties
@@ -1,130 +1,0 @@
-annotation.processing.enabled=true
-annotation.processing.enabled.in.editor=false
-annotation.processing.processors.list=
-annotation.processing.run.all.processors=true
-annotation.processing.source.output=${build.generated.sources.dir}/ap-source-output
-application.title=TrabalhoFinalDesktop
-application.vendor=samue
-build.classes.dir=${build.dir}/classes
-build.classes.excludes=**/*.java,**/*.form
-# This directory is removed when the project is cleaned:
-build.dir=build
-build.generated.dir=${build.dir}/generated
-build.generated.sources.dir=${build.dir}/generated-sources
-# Only compile against the classpath explicitly listed here:
-build.sysclasspath=ignore
-build.test.classes.dir=${build.dir}/test/classes
-build.test.results.dir=${build.dir}/test/results
-compile.on.save=true
-compile.on.save.unsupported.javafx=true
-# Uncomment to specify the preferred debugger connection transport:
-#debug.transport=dt_socket
-debug.classpath=\
-    ${run.classpath}
-debug.modulepath=\
-    ${run.modulepath}
-debug.test.classpath=\
-    ${run.test.classpath}
-debug.test.modulepath=\
-    ${run.test.modulepath}
-# This directory is removed when the project is cleaned:
-dist.dir=dist
-dist.jar=${dist.dir}/TrabalhoFinalDesktop.jar
-dist.javadoc.dir=${dist.dir}/javadoc
-endorsed.classpath=
-excludes=
-file.reference.mysql-connector-java-8.0.18.jar=C:\\Users\\Aluno\\Downloads\\mysql-connector-java-8.0.18\\mysql-connector-java-8.0.18.jar
-includes=**
-# Non-JavaFX jar file creation is deactivated in JavaFX 2.0+ projects
-jar.archive.disabled=true
-jar.compress=false
-javac.classpath=\
-    ${javafx.classpath.extension}:\
-    ${file.reference.mysql-connector-java-8.0.18.jar}
-# Space-separated list of extra javac options
-javac.compilerargs=
-javac.deprecation=false
-javac.external.vm=false
-javac.modulepath=
-javac.processormodulepath=
-javac.processorpath=\
-    ${javac.classpath}
-javac.source=1.8
-javac.target=1.8
-javac.test.classpath=\
-    ${javac.classpath}:\
-    ${build.classes.dir}
-javac.test.modulepath=\
-    ${javac.modulepath}
-javac.test.processorpath=\
-    ${javac.test.classpath}
-javadoc.additionalparam=
-javadoc.author=false
-javadoc.encoding=${source.encoding}
-javadoc.html5=false
-javadoc.noindex=false
-javadoc.nonavbar=false
-javadoc.notree=false
-javadoc.private=false
-javadoc.splitindex=true
-javadoc.use=true
-javadoc.version=false
-javadoc.windowtitle=
-javafx.application.implementation.version=1.0
-javafx.binarycss=false
-javafx.classpath.extension=\
-    ${platforms.JDK_8.home}/jre/lib/javaws.jar:\
-    ${platforms.JDK_8.home}/jre/lib/deploy.jar:\
-    ${platforms.JDK_8.home}/jre/lib/plugin.jar
-javafx.deploy.allowoffline=true
-# If true, application update mode is set to 'background', if false, update mode is set to 'eager'
-javafx.deploy.backgroundupdate=false
-javafx.deploy.embedJNLP=true
-javafx.deploy.includeDT=true
-# Definir como verdadeiro impede a cria\u00e7\u00e3o de c\u00f3pia tempor\u00e1ria de artefatos de implanta\u00e7\u00e3o antes de cada execu\u00e7\u00e3o (desativa as execu\u00e7\u00f5es atuais)
-javafx.disable.concurrent.runs=false
-# Definir como verdadeiro ativa v\u00e1rias execu\u00e7\u00f5es simult\u00e2neas do mesmo WebStart ou projeto Run-in-Browser
-javafx.enable.concurrent.external.runs=false
-# This is a JavaFX project
-javafx.enabled=true
-javafx.fallback.class=com.javafx.main.NoJavaFXFallback
-# Main class for JavaFX
-javafx.main.class=Principal.Main
-javafx.preloader.class=
-# This project does not use Preloader
-javafx.preloader.enabled=false
-javafx.preloader.jar.filename=
-javafx.preloader.jar.path=
-javafx.preloader.project.path=
-javafx.preloader.type=none
-# Set true for GlassFish only. Rebases manifest classpaths of JARs in lib dir. Not usable with signed JARs.
-javafx.rebase.libs=false
-javafx.run.height=600
-javafx.run.width=800
-jlink.launcher=false
-jlink.launcher.name=TrabalhoFinalDesktop
-# Pre-JavaFX 2.0 WebStart is deactivated in JavaFX 2.0+ projects
-jnlp.enabled=false
-# Main class for Java launcher
-main.class=com.javafx.main.Main
-# Para uma melhor seguran\u00e7a, especifique um atributo de manifesto da Base de c\u00f3digo mais restrito para impedir que RIAs sejam adaptados
-manifest.custom.codebase=*
-# Especificar o atributo de manifesto de Permiss\u00f5es para substituir o default (op\u00e7\u00f5es: sandbox, all-permissions)
-manifest.custom.permissions=
-manifest.file=manifest.mf
-meta.inf.dir=${src.dir}/META-INF
-mkdist.disabled=false
-platform.active=JDK_8
-run.classpath=\
-    ${javac.classpath}:\
-    ${build.classes.dir}
-run.modulepath=\
-    ${javac.modulepath}
-run.test.classpath=\
-    ${javac.test.classpath}:\
-    ${build.test.classes.dir}
-run.test.modulepath=\
-    ${javac.test.modulepath}
-source.encoding=UTF-8
-src.dir=src
-test.src.dir=test
