--- diario/Cursos/src/Principal/Main.java
@@ -1,33 +1,0 @@
-
-package Principal;
-
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
-import javafx.stage.Stage;
-
-public class Main extends Application{
-    private static Stage stage;
-    @Override
-    public void start(Stage stage) throws Exception{
-        Parent root = FXMLLoader.load(getClass().getResource("FXML.fxml"));
-        
-        Scene scene = new Scene(root);
-        
-        stage.setScene(scene);
-        stage.show();
-        setStage(stage);
-    }
-    public static void main(String[] args) {
-        launch(args);
-    }
-    
-      public static Stage getStage(){
-        return stage;
-    }
-
-    private void setStage(Stage stage) {
-        this.stage = stage;
-    }
-}
